requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyPDF2
python-docx
pandas
numpy
groq
python-dotenv
tqdm
